news_summarizer/requirements.txt
newspaper3k
konlpy
networkx
scikit-learn
numpy
transformers
torch
flask
rouge
